requirements.txt
Django
Pillow
djangorestframework
markdown
django-filter
django-cors-headers
PyJWT
